requirements.txt
aiohttp==3.8.1
aiomysql==0.1.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
frozenlist==1.3.1
idna==3.3
multidict==6.0.2
mysql-connector-python==8.0.30
PyMySQL==1.0.2
requests==2.28.1
urllib3==1.26.11
yarl==1.8.1
